third_party/blink/web_tests/platform/generic/crypto/subtle/hkdf/unwrapKey-expected.txt

Test unwrapping an HKDF key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS wrappingKey.algorithm.name is "AES-GCM"
PASS typeof wrappingKey.extractable is 'boolean'
PASS wrappingKey.extractable is false
PASS wrappingKey.usages.join(',') is "encrypt,unwrapKey"
PASS unwrappedHkdfKey.algorithm.name is "HKDF"
PASS typeof unwrappedHkdfKey.extractable is 'boolean'
PASS unwrappedHkdfKey.extractable is false
PASS unwrappedHkdfKey.usages.join(',') is "deriveBits"

Unwrap an HKDF key using pkcs8 as the format.
error is: NotSupportedError: Unsupported import key format for algorithm
PASS successfullyParsed is true

TEST COMPLETE
